gpt/requirements.txt
matplotlib==3.7.1
numpy==1.24.3
openai==0.27.7
pandas==2.0.2
python-dotenv==1.0.0
Requests==2.31.0
seaborn==0.12.2
